requirements.txt
affine==2.3.0
anyio @ file:///opt/concourse/worker/volumes/live/4ff95164-2ca6-4efb-5c02-5afee907620d/volume/anyio_1617783322708/work/dist
appdirs==1.4.4
appnope @ file:///opt/concourse/worker/volumes/live/4f734db2-9ca8-4d8b-5b29-6ca15b4b4772/volume/appnope_1606859466979/work
argon2-cffi @ file:///opt/concourse/worker/volumes/live/4afd07c8-7fc3-4a09-6326-d8c70269eb33/volume/argon2-cffi_1613037490059/work
asciitree==0.3.3
asgiref==3.2.10
astroid==2.5.3
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
autopep8==1.5.7
Babel @ file:///tmp/build/80754af9/babel_1607110387436/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
bqplot==0.12.23
branca==0.4.2
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///opt/concourse/worker/volumes/live/ca3c0fea-9d0f-4489-6889-89b42078ed37/volume/cffi_1613246931142/work
cftime==1.4.1
chardet @ file:///opt/concourse/worker/volumes/live/9efbf151-b45b-463d-6340-a5c399bf00b7/volume/chardet_1607706825988/work
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cloudpickle==1.6.0
conda==4.9.2
conda-package-handling @ file:///opt/concourse/worker/volumes/live/41a2ba95-d84c-4c06-4d6e-5665866bc5b4/volume/conda-package-handling_1603018173200/work
constructor @ file:///opt/concourse/worker/volumes/live/71329617-32f2-486d-789c-015aca045556/volume/constructor_1598032202984/work
cryptography @ file:///opt/concourse/worker/volumes/live/2a751272-b8a2-45c1-4dc6-ec2719cf7283/volume/cryptography_1613401428693/work
cycler==0.10.0
dask==2021.3.1
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///home/ktietz/src/ci_mi/defusedxml_1612808095953/work
Django==3.1
django-environ==0.4.5
django-sortedm2m==3.0.2
djangorestframework==3.12.2
earthpy==0.9.2
entrypoints==0.3
fasteners==0.16
Fiona==1.8.19
flake8==3.8.4
fsspec==0.8.7
geopandas==0.9.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imageio==2.9.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
intake==0.6.2
intake-stac @ git+https://github.com/intake/intake-stac@3db971897dac3680504b37ec259f64f424163a81
intake-xarray==0.5.0
ipykernel @ file:///opt/concourse/worker/volumes/live/73e8766c-12c3-4f76-62a6-3dea9a7da5b7/volume/ipykernel_1596206701501/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipyleaflet==0.13.6
ipython==7.22.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.3
isort==5.8.0
jedi @ file:///opt/concourse/worker/volumes/live/5006b7b5-a924-4788-6cfe-ae05d8be8830/volume/jedi_1606932947370/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
json5==0.9.5
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///opt/concourse/worker/volumes/live/a699b83f-e941-4170-5136-bf87e3f37756/volume/jupyter_core_1612213304212/work
jupyter-packaging @ file:///tmp/build/80754af9/jupyter-packaging_1613502826984/work
jupyter-server @ file:///opt/concourse/worker/volumes/live/8c909028-d80d-4303-5883-52f907a3cf74/volume/jupyter_server_1616084052708/work
jupyterlab @ file:///tmp/build/80754af9/jupyterlab_1617303052588/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1617134334258/work
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
locket==0.2.1
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.2
munch==2.5.0
nbclassic @ file:///tmp/build/80754af9/nbclassic_1616085367084/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///opt/concourse/worker/volumes/live/d4b0787b-b6c8-4d28-5453-3381885d5b33/volume/nbconvert_1601914848300/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
netCDF4==1.5.6
networkx==2.5.1
notebook @ file:///opt/concourse/worker/volumes/live/918fd277-9373-4d2e-5e16-cbd1adfac363/volume/notebook_1611341013189/work
numcodecs==0.7.3
numpy==1.20.1
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.2.2
pandocfilters @ file:///opt/concourse/worker/volumes/live/315ac9bb-93fd-4adc-6795-345692fcfaed/volume/pandocfilters_1605120446899/work
parso==0.7.0
partd==1.1.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.1.2
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycodestyle==2.7.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyflakes==2.2.0
Pygments @ file:///tmp/build/80754af9/pygments_1613849567472/work
pylint==2.7.4
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyproj==3.0.1
pyrsistent @ file:///opt/concourse/worker/volumes/live/656e0c1b-ef87-4251-4a51-1290b2351993/volume/pyrsistent_1600141745371/work
PySocks @ file:///opt/concourse/worker/volumes/live/ef943889-94fc-4539-798d-461c60b77804/volume/pysocks_1605305801690/work
pystac==0.5.6
python-dateutil==2.7.5
python-decouple==3.3
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
rasterio==1.2.2
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
ruamel-yaml==0.15.87
sat-search==0.3.0
sat-stac==0.4.1
scikit-image==0.15.0
scipy==1.6.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
Shapely==1.7.1
six @ file:///opt/concourse/worker/volumes/live/f983ba11-c9fe-4dff-7ce7-d89b95b09771/volume/six_1605205318156/work
sniffio @ file:///opt/concourse/worker/volumes/live/838fa2d9-a35b-4591-50ce-1f1a39baa1df/volume/sniffio_1614030463440/work
snuggs==1.4.7
sqlparse==0.4.1
terminado==0.9.2
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
tifffile==2021.4.8
toml==0.10.2
toolz==0.11.1
tornado @ file:///opt/concourse/worker/volumes/live/d531d395-893c-4ca1-6a5f-717b318eb08c/volume/tornado_1606942307627/work
tqdm @ file:///tmp/build/80754af9/tqdm_1611857934208/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
traittypes==0.2.1
typed-ast==1.4.3
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1611694770489/work
voila==0.2.7
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.17.0
zarr==2.7.0
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
